backend/grok-voice-mem0-demo/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-dotenv==1.0.0
mem0ai==1.0.1
httpx==0.25.0
